src/crypto/pqc.py
import base64
import os
from src.utils.logger import setup_logger


class QuantumCrypto:
    def __init__(self):
        self.logger = setup_logger("QuantumCrypto")

    def generate_kem_keypair(self):
        """Generate key pair for key exchange"""
        try:
            # Placeholder for quantum key generation
            public_key = b"quantum_public_key_placeholder"
            return public_key, "kem_object_placeholder"
        except Exception as e:
            self.logger.error(f"Failed to generate KEM keypair: {e}")
            return None, None

    def key_exchange(self, peer_public_key, my_kem):
        """Perform key exchange"""
        try:
            # Placeholder for quantum key exchange
            ciphertext = b"quantum_ciphertext_placeholder"
            shared_secret = b"quantum_shared_secret_placeholder"
            return ciphertext, shared_secret
        except Exception as e:
            self.logger.error(f"Key exchange failed: {e}")
            return None, None

    def generate_signing_keypair(self):
        """Generate key pair for signatures"""
        try:
            # Placeholder for quantum signature generation
            public_key = b"quantum_sig_public_key_placeholder"
            return public_key, "signer_object_placeholder"
        except Exception as e:
            self.logger.error(f"Failed to generate signing keypair: {e}")
            return None, None

    def sign_message(self, message, signer):
        """Sign a message"""
        try:
            if isinstance(message, str):
                message = message.encode('utf-8')
            return b"quantum_signature_placeholder"
        except Exception as e:
            self.logger.error(f"Failed to sign message: {e}")
            return None

    def verify_signature(self, message, signature, public_key):
        """Verify a signature"""
        try:
            if isinstance(message, str):
                message = message.encode('utf-8')
            return True  # Placeholder
        except Exception as e:
            self.logger.error(f"Signature verification failed: {e}")
            return False

    def encrypt_message(self, message, shared_secret):
        """Encrypt message using shared secret"""
        try:
            if isinstance(message, str):
                message = message.encode('utf-8')

            # Simple XOR encryption (placeholder for quantum encryption)
            encrypted = bytearray()
            for i, char in enumerate(message):
                encrypted.append(char ^ shared_secret[i % len(shared_secret)])

            return base64.b64encode(encrypted).decode('utf-8')

        except Exception as e:
            self.logger.error(f"Encryption failed: {e}")
            return None

    def decrypt_message(self, encrypted_data, shared_secret):
        """Decrypt message using shared secret"""
        try:
            encrypted_data = base64.b64decode(encrypted_data)

            # Simple XOR decryption (placeholder for quantum decryption)
            decrypted = bytearray()
            for i, char in enumerate(encrypted_data):
                decrypted.append(char ^ shared_secret[i % len(shared_secret)])

            return decrypted.decode('utf-8')

        except Exception as e:
            self.logger.error(f"Decryption failed: {e}")
            return None

    def get_supported_algorithms(self):
        """Get list of supported quantum algorithms"""
        try:
            # Try to import liboqs if available
            import liboqs
            kems = liboqs.get_enabled_KEM_mechanisms()
            sigs = liboqs.get_enabled_sig_mechanisms()
            return {
                'key_encapsulation': kems,
                'signatures': sigs
            }
        except ImportError:
            # Return placeholder if liboqs not available
            return {
                'key_encapsulation': ['Kyber512', 'Kyber768', 'Kyber1024'],
                'signatures': ['Dilithium2', 'Dilithium3', 'Dilithium5']
            }


def test_quantum_crypto():
    """Test quantum cryptography functionality"""
    crypto = QuantumCrypto()

    print("🔐 Testing Quantum Cryptography...")

    # Test encryption/decryption with placeholder
    shared_secret = b"test_shared_secret"
    test_message = "Hello Quantum World!"

    encrypted = crypto.encrypt_message(test_message, shared_secret)
    decrypted = crypto.decrypt_message(encrypted, shared_secret)

    if decrypted == test_message:
        print("✅ Basic encryption/decryption working")
    else:
        print("❌ Basic encryption/decryption failed")

    # Test algorithm detection
    algorithms = crypto.get_supported_algorithms()
    print(
        f"✅ Supported algorithms detected: {len(algorithms['key_encapsulation'])} KEMs, {len(algorithms['signatures'])} signatures")

    print("🎉 Quantum crypto tests completed!")
    return True


if __name__ == "__main__":
    test_quantum_crypto()