requirements.txt
# Essential dependencies
qrcode[pil]>=7.4.2
Pillow>=10.0.0
opencv-python>=4.8.0
stegano>=0.9.3

# Optional dependencies
liboqs-python>=0.8.0
cryptography>=41.0.0
websockets>=11.0.0

# Development dependencies
pytest>=7.0.0